Move service worker to public folder
