+++ public/serviceWorker.ts
