+++ tsconfig.json
@@ -21,6 +21,6 @@
     "jsx": "react"
   },
   "include": [
-    "src"
+    "src", "public/serviceWorker.ts"
   ]
 }
